added centralize shade mocks
