+++ apps/posts/test/setup.ts
@@ -1,17 +1,5 @@
 import '@testing-library/jest-dom';
-import {vi} from 'vitest';
+import {setupShadeMocks} from '@tryghost/admin-x-framework/test/setup';
 
-// Mock window.matchMedia
-Object.defineProperty(window, 'matchMedia', {
-    writable: true,
-    value: vi.fn().mockImplementation(query => ({
-        matches: false,
-        media: query,
-        onchange: null,
-        addListener: vi.fn(), // deprecated
-        removeListener: vi.fn(), // deprecated
-        addEventListener: vi.fn(),
-        removeEventListener: vi.fn(),
-        dispatchEvent: vi.fn()
-    }))
-}); 
+// Set up common mocks for shade components
+setupShadeMocks();
